+++ src/RoadRegistry.BackOffice.Api/Information/InformationController.cs
@@ -4,7 +4,6 @@ using Be.Vlaanderen.Basisregisters.Api;
 using Infrastructure;
 using Infrastructure.Authentication;
 using Infrastructure.Authorization;
-using Infrastructure.Controllers.Attributes;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -13,8 +12,7 @@ using Microsoft.AspNetCore.Mvc;
 [AdvertiseApiVersions(Version.CurrentAdvertised)]
 [ApiRoute("information")]
 [ApiExplorerSettings(GroupName = "Info")]
-[ApiKey]
-[Authorize(AuthenticationSchemes = AuthenticationSchemes.AllBearerSchemes, Policy = AcmIdmConstants.PolicyNames.VoInfo)]
+[Authorize(AuthenticationSchemes = AuthenticationSchemes.AllSchemes, Policy = AcmIdmConstants.PolicyNames.VoInfo)]
 public partial class InformationController : ApiController
 {
     private readonly IMediator _mediator;