+++ src/RoadRegistry.BackOffice.Api/Uploads/UploadController.cs
@@ -1,7 +1,5 @@
 namespace RoadRegistry.BackOffice.Api.Uploads;
 
-using System;
-using System.Threading.Tasks;
 using Abstractions.Exceptions;
 using BackOffice.Extracts;
 using Be.Vlaanderen.Basisregisters.AcmIdm;
@@ -13,19 +11,19 @@ using FluentValidation.Results;
 using Hosts.Infrastructure.Options;
 using Infrastructure.Authentication;
 using Infrastructure.Controllers;
-using Infrastructure.Controllers.Attributes;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System;
+using System.Threading.Tasks;
 using Version = Infrastructure.Version;
 
 [ApiVersion(Version.Current)]
 [AdvertiseApiVersions(Version.CurrentAdvertised)]
 [ApiRoute("upload")]
 [ApiExplorerSettings(GroupName = "Upload")]
-[ApiKey]
-[Authorize(AuthenticationSchemes = AuthenticationSchemes.AllBearerSchemes, Policy = PolicyNames.IngemetenWeg.Beheerder)]
+[Authorize(AuthenticationSchemes = AuthenticationSchemes.AllSchemes, Policy = PolicyNames.IngemetenWeg.Beheerder)]
 public partial class UploadController : BackofficeApiController
 {
     private readonly IMediator _mediator;