+++ src/RoadRegistry.BackOffice/Core/IRoadNetworks.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 public interface IRoadNetworks
 {
     Task<RoadNetwork> Get(CancellationToken cancellationToken);
-    Task<RoadNetwork> Get(bool restoreSnapshot, ProcessMessageHandler processMessage, CancellationToken cancellationToken);
+    Task<RoadNetwork> Get(bool restoreSnapshot, ProcessMessageHandler cancelMessageProcessing, CancellationToken cancellationToken);
     Task<(RoadNetwork, int)> GetWithVersion(CancellationToken cancellationToken);
-    Task<(RoadNetwork, int)> GetWithVersion(bool restoreSnapshot, ProcessMessageHandler processMessage, CancellationToken cancellationToken);
+    Task<(RoadNetwork, int)> GetWithVersion(bool restoreSnapshot, ProcessMessageHandler cancelMessageProcessing, CancellationToken cancellationToken);
 }